+++ internal/fs/inode/dir.go
@@ -244,7 +244,7 @@ func NewDirInode(
 	bucket *gcsx.SyncerBucket,
 	mtimeClock timeutil.Clock,
 	cacheClock timeutil.Clock,
-	typeCacheMaxSizeMB int,
+	typeCacheMaxSizeMB int64,
 	isHNSEnabled bool,
 ) (d DirInode) {
