+++ internal/fs/inode/dir_test.go
@@ -98,7 +98,7 @@ func (t *DirTest) resetInode(implicitDirs, enableNonexistentTypeCache, enableMan
 	t.resetInodeWithTypeCacheConfigs(implicitDirs, enableNonexistentTypeCache, enableManagedFoldersListing, cfg.DefaultTypeCacheMaxSizeMB, typeCacheTTL)
 }
 
-func (t *DirTest) resetInodeWithTypeCacheConfigs(implicitDirs, enableNonexistentTypeCache, enableManagedFoldersListing bool, typeCacheMaxSizeMB int, typeCacheTTL time.Duration) {
+func (t *DirTest) resetInodeWithTypeCacheConfigs(implicitDirs, enableNonexistentTypeCache, enableManagedFoldersListing bool, typeCacheMaxSizeMB int64, typeCacheTTL time.Duration) {
 	if t.in != nil {
 		t.in.Unlock()
 	}
@@ -656,7 +656,7 @@ func (t *DirTest) LookUpChild_NonExistentTypeCache_ImplicitDirsEnabled() {
 
 func (t *DirTest) LookUpChild_TypeCacheEnabled() {
 	inputs := []struct {
-		typeCacheMaxSizeMB int
+		typeCacheMaxSizeMB int64
 		typeCacheTTL       time.Duration
 	}{{
 		typeCacheMaxSizeMB: cfg.DefaultTypeCacheMaxSizeMB,
@@ -689,7 +689,7 @@ func (t *DirTest) LookUpChild_TypeCacheEnabled() {
 
 func (t *DirTest) LookUpChild_TypeCacheDisabled() {
 	inputs := []struct {
-		typeCacheMaxSizeMB int
+		typeCacheMaxSizeMB int64
 		typeCacheTTL       time.Duration
 	}{{
 		typeCacheMaxSizeMB: 0,
